refactor verify codesigned
